requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0  # Optional, for advanced analytics
matplotlib>=3.7.0  # Optional, for visualizations
seaborn>=0.12.0  # Optional, for enhanced visualizations